requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.1
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
